+++ services/payment-service/src/main/java/com/sj/ecommerce/payment_service/dto/PaymentResponse.java
@@ -1,6 +1,5 @@
 package com.sj.ecommerce.payment_service.dto;
 
-import java.math.BigDecimal;
 import java.time.Instant;
 
-public record PaymentResponse(Long id, Long orderId, Long userId, BigDecimal amount, String idempotencyKey, String status, Instant createdAt) {}
+public record PaymentResponse(Long id, Long orderId, Long userId, Double amount, String idempotencyKey, String status, Instant createdAt) {}
